# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/search/movie": {
            "get": {
                "description": "Search for movies using The Movie Database (TMDB) API. Provide search terms via query parameter to find matching movies.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Search movies via TMDB",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query for movie titles",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TMDB search results containing matching movies",
                        "schema": {
                            "$ref": "#/definitions/tmdb.MovieResult"
                        }
                    },
                    "400": {
                        "description": "Missing or empty query parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "TMDB API error or server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/watched": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Get all watched movies",
                "responses": {
                    "200": {
                        "description": "List of all watched movies (array may be empty)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Movie"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error while fetching watched movies",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Mark a movie as watched by providing a TMDB ID. If the movie doesn't exist in the database, it will be fetched from TMDB and created. If no date is provided, the current date is used.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Mark a movie as watched",
                "parameters": [
                    {
                        "description": "Watched movie data (must include `id` as TMDB ID; optional `date` in YYYY-MM-DD format)",
                        "name": "watched",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Watched"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully marked movie as watched",
                        "schema": {
                            "$ref": "#/definitions/model.Watched"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing required `id` field",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error, database error, or TMDB API error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Movie": {
            "type": "object",
            "properties": {
                "budget": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imdb_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "original_language": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "release": {
                    "type": "string"
                },
                "revenue": {
                    "type": "integer"
                },
                "runtime": {
                    "type": "integer"
                },
                "vote_average": {
                    "type": "number"
                },
                "watched_date": {
                    "type": "string"
                }
            }
        },
        "model.Watched": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "tmdb.MovieResult": {
            "type": "object",
            "properties": {
                "adult": {
                    "type": "boolean"
                },
                "backdrop_path": {
                    "type": "string"
                },
                "genre_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "original_language": {
                    "type": "string"
                },
                "original_title": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "popularity": {
                    "type": "number"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "video": {
                    "type": "boolean"
                },
                "vote_average": {
                    "type": "number"
                },
                "vote_count": {
                    "type": "integer"
                }
            }
        }
    }
}